final new project structure
