--- ATM/src/main/java/ru/sber/javaschool/balance/BalanceNotOverdraft.java
@@ -1,16 +1,0 @@
-package ru.sber.javaschool.balance;
-
-import java.math.BigDecimal;
-
-public class BalanceNotOverdraft extends Balance {
-    public BalanceNotOverdraft(Currency currency, BigDecimal amount) {
-        super(currency, amount);
-    }
-
-    @Override
-    public void setAmount(BigDecimal amount) {
-        if (amount.compareTo(BigDecimal.ZERO) < 0)
-            throw new BalanceException("Баланс не может быть меньше 0");
-        super.setAmount(amount);
-    }
-}
