--- ATM/src/main/java/ru/sber/javaschool/balance/Currency.java
@@ -1,7 +1,0 @@
-package ru.sber.javaschool.balance;
-
-public enum Currency {
-    RUR,
-    USD,
-    EUR
-}
