--- ATM/src/main/java/ru/sber/javaschool/client/Client.java
@@ -1,14 +1,0 @@
-package ru.sber.javaschool.client;
-
-import ru.sber.javaschool.account.Account;
-
-import java.util.List;
-
-public class Client {
-    /** Поле номера клиента */
-    private int clientNum;
-    /** Поле персональной информации */
-    private Person person;
-    /** Коллекция счетов клиента */
-    private List<Account<?>> accounts;
-}
