--- ATM/src/main/java/ru/sber/javaschool/client/Person.java
@@ -1,17 +1,0 @@
-package ru.sber.javaschool.client;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class Person {
-    /** Поле имени */
-    private String firstName;
-    /** Поле отчества */
-    private String middleName;
-    /** Поле фамилии */
-    private String lastName;
-    /** Поле возраста */
-    private int age;
-}
