--- ATM/src/main/java/ru/sber/javaschool/terminal/CardReader.java
@@ -1,22 +1,0 @@
-package ru.sber.javaschool.terminal;
-
-/**
- * Интерфейс карт-ридера
- */
-public interface CardReader {
-    /**
-     * Принять карту
-     */
-    void inputCard();
-
-    /**
-     * Извлечь карту
-     */
-    void extractCard();
-
-    /**
-     * Проверить наличие карты
-     * @return возвращает признак наличия карты в карт-ридере
-     */
-    boolean checkCardAvailability();
-}
