--- ATM/src/main/java/ru/sber/javaschool/user/User.java
@@ -1,22 +1,0 @@
-package ru.sber.javaschool.user;
-
-import ru.sber.javaschool.card.Card;
-import ru.sber.javaschool.terminal.Terminal;
-
-/**
- * Интерфейс пользователя
- */
-public interface User {
-    /**
-     * Вставить карту в терминал
-     * @param terminal - терминал
-     */
-    void insertCard(Terminal terminal);
-
-    /**
-     * Запросить баланс в терминале
-     * @param terminal - терминал
-     */
-    void requestBalance(Terminal terminal);
-    void setCard(Card card);
-}
