+++ server/src/main/java/ru/burlutskiysergey/javaschool/balance/Balance.java
@@ -1,0 +1,18 @@
+package ru.burlutskiysergey.javaschool.balance;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.math.BigDecimal;
+
+@Getter
+@AllArgsConstructor
+public abstract class Balance {
+    /** Поле валюты */
+    private Currency currency;
+
+    /** Поле суммы */
+    @Setter
+    private BigDecimal amount;
+}
