+++ server/src/main/java/ru/burlutskiysergey/javaschool/balance/BalanceException.java
@@ -1,0 +1,23 @@
+package ru.burlutskiysergey.javaschool.balance;
+
+public class BalanceException extends RuntimeException {
+    public BalanceException() {
+        super();
+    }
+
+    public BalanceException(String message) {
+        super(message);
+    }
+
+    public BalanceException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public BalanceException(Throwable cause) {
+        super(cause);
+    }
+
+    protected BalanceException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
