+++ server/src/main/java/ru/burlutskiysergey/javaschool/balance/BalanceNotOverdraft.java
@@ -1,0 +1,16 @@
+package ru.burlutskiysergey.javaschool.balance;
+
+import java.math.BigDecimal;
+
+public class BalanceNotOverdraft extends Balance {
+    public BalanceNotOverdraft(Currency currency, BigDecimal amount) {
+        super(currency, amount);
+    }
+
+    @Override
+    public void setAmount(BigDecimal amount) {
+        if (amount.compareTo(BigDecimal.ZERO) < 0)
+            throw new BalanceException("Баланс не может быть меньше 0");
+        super.setAmount(amount);
+    }
+}
