+++ server/src/main/java/ru/burlutskiysergey/javaschool/balance/Currency.java
@@ -1,0 +1,7 @@
+package ru.burlutskiysergey.javaschool.balance;
+
+public enum Currency {
+    RUR,
+    USD,
+    EUR
+}
