+++ server/src/main/java/ru/burlutskiysergey/javaschool/card/Card.java
@@ -1,0 +1,29 @@
+package ru.burlutskiysergey.javaschool.card;
+
+import java.math.BigDecimal;
+
+/**
+ * Интерфейс карты
+ */
+public interface Card {
+    /**
+     * Проверить пин-код карты
+     * @param pinCode - введённый пин-код
+     * @return возвращает признак правильно ли введён пин-код
+     */
+    boolean checkPin(int pinCode);
+
+    /**
+     * Получить баланс карты
+     * @return возвращает баланс карты
+     */
+    BigDecimal getBalance();
+
+    /**
+     * Установить пин-код карты
+     * @param pinCode - пин-код
+     */
+    void setPinCode(int pinCode);
+
+    void setBalance(BigDecimal balance);
+}
