+++ server/src/main/java/ru/burlutskiysergey/javaschool/client/Client.java
@@ -1,0 +1,14 @@
+package ru.burlutskiysergey.javaschool.client;
+
+import ru.burlutskiysergey.javaschool.account.Account;
+
+import java.util.List;
+
+public class Client {
+    /** Поле номера клиента */
+    private int clientNum;
+    /** Поле персональной информации */
+    private Person person;
+    /** Коллекция счетов клиента */
+    private List<Account<?>> accounts;
+}
