+++ server/src/main/java/ru/burlutskiysergey/javaschool/client/Person.java
@@ -1,0 +1,17 @@
+package ru.burlutskiysergey.javaschool.client;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class Person {
+    /** Поле имени */
+    private String firstName;
+    /** Поле отчества */
+    private String middleName;
+    /** Поле фамилии */
+    private String lastName;
+    /** Поле возраста */
+    private int age;
+}
