+++ server/src/main/java/ru/burlutskiysergey/javaschool/terminal/CardReader.java
@@ -1,0 +1,22 @@
+package ru.burlutskiysergey.javaschool.terminal;
+
+/**
+ * Интерфейс карт-ридера
+ */
+public interface CardReader {
+    /**
+     * Принять карту
+     */
+    void inputCard();
+
+    /**
+     * Извлечь карту
+     */
+    void extractCard();
+
+    /**
+     * Проверить наличие карты
+     * @return возвращает признак наличия карты в карт-ридере
+     */
+    boolean checkCardAvailability();
+}
