--- src/.gitignore
@@ -1,3 +1,0 @@
-.idea
-target
-*.iml
